+++ src/main/java/pe/edu/upc/greenly/repositories/DonacionRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Donacion;
 
-public interface DonacionRepository extends JpaRepository<Donacion, Integer> {
+public interface DonacionRepository extends JpaRepository<Donacion, Long> {
 }
